fix(refactor): move utils to common folder
