+++ apps/api/src/app/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { PrismaModule } from '../prisma/prisma.module';
-import { UtilsModule } from '../../utils/utils.module';
+import { UtilsModule } from '../../common/utils/utils.module';
 
 @Module({
   imports: [PrismaModule, UtilsModule],
