+++ apps/api/src/common/utils/utils.service.ts
